# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nutri/patients": {
      "get": {
        "description": "Nome + ciclo atual (o aberto; se não houver, o fechado mais recente; senão null). Ordem por nome, desempate por id. Nenhuma métrica é calculada aqui — adesão e relatório vivem nas rotas do paciente. Limite v0: a credencial stub dá o papel \"nutri do sistema\", então a listagem não é escopada por nutri responsável (entra com a auth real).",
        "operationId": "PatientsController_listar",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "lista de pacientes + ciclo atual"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          }
        },
        "summary": "Listar pacientes com o ciclo atual de cada um",
        "tags": [
          "Pacientes (só nutri)"
        ]
      },
      "post": {
        "description": "Cria o paciente vinculado à nutricionista responsável e devolve-o na MESMA forma do item da listagem (cicloAtual null), para o cliente não precisar de uma segunda chamada. Coleta mínima: só o nome (LGPD) — e-mail, telefone, peso e altura não são coletados porque nada os consome. Escreve apenas `patient`: sem plano, sem ciclo, sem programação. Plano entra por outra via (import de PDF, Fase 4).",
        "operationId": "PatientsController_criar",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Ana Ribeiro"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "paciente criado"
          },
          "400": {
            "description": "name ausente, vazio, não-string ou acima de 120 caracteres"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "422": {
            "description": "nenhuma nutricionista cadastrada, ou mais de uma (a credencial stub não distingue a responsável)"
          }
        },
        "summary": "Cadastrar um paciente",
        "tags": [
          "Pacientes (só nutri)"
        ]
      }
    },
    "/nutri/patients/{patientId}": {
      "get": {
        "description": "Nome, contato, peso, altura e nível de exposição — o que o formulário de edição precisa preencher. Sem ciclo: a regra de \"qual ciclo mostrar\" mora na listagem (015/D2) e em nenhum outro lugar.",
        "operationId": "PatientsController_detalhe",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a ficha do paciente"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "paciente não encontrado"
          }
        },
        "summary": "A ficha de um paciente",
        "tags": [
          "Pacientes (só nutri)"
        ]
      },
      "patch": {
        "description": "PATCH parcial: campo AUSENTE do corpo preserva o valor atual; campo com `null` LIMPA (apagar dado de saúde é um direito, LGPD). Corpo sem nenhum campo conhecido é no-op, não erro. Campos: name, email, phone, heightCm, weightKg, exposure (hidden|percent|macros|full_kcal).",
        "operationId": "PatientsController_atualizar",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Ana Ribeiro",
                  "email": "ana@exemplo.com",
                  "phone": "11999990000",
                  "heightCm": 165,
                  "weightKg": 62.5,
                  "exposure": "macros"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "a ficha depois da alteração"
          },
          "400": {
            "description": "campo com formato ou faixa inválidos"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "paciente não encontrado"
          }
        },
        "summary": "Editar a ficha do paciente",
        "tags": [
          "Pacientes (só nutri)"
        ]
      },
      "delete": {
        "description": "Cascata PARA BAIXO: planos e o grafo inteiro (tipos-de-dia, refeições, opções, itens, programação da semana), ciclos e vigências. RECUSA com 409 se houver registro de refeição — histórico de saúde não é apagado por exclusão de cadastro.",
        "operationId": "PatientsController_excluir",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "paciente e grafo excluídos"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "paciente não encontrado"
          },
          "409": {
            "description": "o paciente tem registro de refeição"
          }
        },
        "summary": "Excluir o paciente",
        "tags": [
          "Pacientes (só nutri)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bamboo API",
    "description": "API da alça do paciente — ver \"o agora\" (US1) e substituir um alimento dentro do grupo, com quantidade recalculada e medida caseira (US2). v0: auth stub (paciente fixo por env).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}